editoast: authz: improve testing utilities
